Adjust formatting of NEWS file
All previous entries in NEWS have a format of headings following nested
levels of indentation — the release is unindented, then subheadings with
a 2-space indentation, then the contents of each subhead are blocks
indented with 2 spaces.

In the 2021b release, there is some commentary about the release that
isn't indented to the level of the "Briefly" sub-block, but also doesn't
have its own heading. I think it would be preferable to put this note
under a heading rather than as free-floating content at the level of other
subheads.